Add factory around importer that takes args, rename to makeFsImporter
